Updates to theme
