--- wp-content/themes/wpinabox-theme/lib/acf/index.php
@@ -1,5 +1,0 @@
-<?php
-
-// Silence is golden.
-
-?>
