--- wp-content/themes/wpinabox-theme/lib/admin.php
@@ -1,16 +1,0 @@
-<?php
-/**
- * Remove extra admin menu items
- *
- * https://isabelcastillo.com/remove-menu-items-wordpress-dashboard
- */
-function wpb_remove_menus() {
-
-	// Comments
-	// remove_menu_page( 'edit-comments.php' );
-}
-
-add_action( 'admin_menu', 'wpb_remove_menus', 999 ); 
-
-
-?>
