--- wp-content/themes/wpinabox-theme/lib/post-types/custom.php
@@ -1,47 +1,0 @@
-<?php
-/*
-class WPB\CustomPost extends TimberPost {
-
-	static function register() {
-
-		$labels = array(
-			'name' => __( 'CustomPosts', 'wpb' ), 
-			'singular_name' => __( 'CustomPost', 'wpb' ),
-			'all_items' => __( 'All CustomPosts', 'wpb' ),
-			'add_new' => __( 'Add New', 'wpb' ),
-			'add_new_item' => __( 'Add New CustomPost', 'wpb' ),
-			'edit' => __( 'Edit', 'wpb' ),
-			'edit_item' => __( 'Edit CustomPost', 'wpb' ),
-			'new_item' => __( 'New CustomPost', 'wpb' ),
-			'view_item' => __( 'View CustomPost', 'wpb' ),
-			'search_items' => __( 'Search CustomPosts', 'wpb' ),
-			'not_found' =>  __( 'Nothing found in the Database.', 'wpb' ),
-			'not_found_in_trash' => __( 'Nothing found in Trash', 'wpb' ),
-			'parent_item_colon' => ''
-		);
-
-		$options = array(
-			'labels' => $labels,
-			'description' => __( 'CustomPosts', 'wpb' ),
-			'public' => true,
-			'publicly_queryable' => true,
-			'exclude_from_search' => false,
-			'show_ui' => true,
-			'query_var' => true,
-			'menu_position' => 5,
-			'rewrite'	=> array( 'slug' => 'custom-post', 'with_front' => false ),
-			'has_archive' => false,
-			'capability_type' => 'post',
-			'hierarchical' => false,
-			'supports' => array( 'title', 'editor', 'revisions')
-		);
-
-		// http://codex.wordpress.org/Function_Reference/register_post_type
-		register_post_type( 'custom_post', $options);			
-
-	}
-
-}
-*/
-
-?>
