--- wp-content/themes/wpinabox-theme/lib/post-types/page.php
@@ -1,14 +1,0 @@
-<?php
-/*
-namespace WPB;
-
-use Timber\Post as TimberPost;
-
-class Page extends TimberPost {
-
-	static function register() {
-	}
-
-}
-*/
-?>
