+++ wp-content/themes/wpinabox-theme/tpl/shortcodes/example.twig
@@ -1,0 +1,4 @@
+<section>
+  <h2>Example Shortcode</h2>
+  <p>{{ content }}</p>
+</section>
